Delete function

+++ books/del.php
@@ -1,36 +1,32 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Dashboard</title>
-        <link rel="stylesheet" href="../static/css/main.css">
-    <link rel="stylesheet" href="../static/css/styles.css">
-    <link rel="stylesheet" href="../static/css/user-books.css">
-    <script src="https://unpkg.com/ionicons@5.2.3/dist/ionicons.js"></script>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-     <script src="https://unpkg.com/sweetalert/dist/sweetalert.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@10"></script>
-
-</head>
-
-
 <?php
 
     
     require_once "../connection.php";
     $book_id = $_GET['id'];
+    $sql = "SELECT * FROM book_view WHERE book_id=$book_id AND is_available=0";
+    $result = mysqli_query($link, $sql);
+    if(mysqli_num_rows($result)==0){
     $sql = "DELETE FROM book WHERE book_id=$book_id";
-    
     if(mysqli_query($link, $sql)){
+        $code = 200;
+        $text = 'OK';
+        $protocol = (isset($_SERVER['SERVER_PROTOCOL']) ? $_SERVER['SERVER_PROTOCOL'] : 'HTTP/1.0');
+        header($protocol . ' ' . $code . ' ' . $text);
+        $GLOBALS['http_response_code'] = $code;
+        echo 'SUCCESS';
         
-            
-            header("location: user_books.php");
-        
+        } else {
+            echo mysqli_error($link);
         } 
+    }
      else{
          
-        
+        $code = 500;
+        $text = 'NOT OK';
+        $protocol = (isset($_SERVER['SERVER_PROTOCOL']) ? $_SERVER['SERVER_PROTOCOL'] : 'HTTP/1.0');
+        header($protocol . ' ' . $code . ' ' . $text);
+        $GLOBALS['http_response_code'] = $code;
+        echo 'Error';
             echo "Oops! Something went wrong. Please try again later.". mysqli_error($link);
         }
